requirements.txt
streamlit==1.35.0
requests==2.31.0
Pillow==10.2.0
PyMuPDF==1.23.26
